10.1: SendGrid webhook + invitation email dispatch + ACCESS_TOKEN_EXPIRE_HOURS
Three deferred Sprint 9 items in one PR. All backend, all extend the
existing tests/integration/ + tests/api/ patterns.

1. POST /webhooks/sendgrid event handler (api/routers/webhooks.py).
   ECDSA P-256 / SHA-256 signature verification (fail-closed: rejects
   when SENDGRID_WEBHOOK_PUBLIC_KEY is unset). Maps SendGrid event
   types to NotificationStatus (delivered → DELIVERED, bounce →
   BOUNCED, etc.), updates Notification.{status, delivered_at,
   opened_at, clicked_at, error_message}, appends DeliveryLog row per
   event. BackgroundTasks dispatch so SendGrid doesn't see slow DB
   writes. Mounted at root (not /api/v1) for external services.
   Registered in api/main.py.

2. send_invitation_email refresh + wiring (api/services/email_service.py
   + api/routers/invitations.py). Existing send_invitation_email used
   a stale URL path (/accept-invitation vs mobile router's /invitation)
   and had no signupflow:// mobile deep link. Refreshed to mirror
   send_password_reset_email: HTML-escape admin_name + org_name (same
   threat model as #78 P1), signupflow://invitation?token=... deep
   link + web fallback. Wired into create_invitation (was a no-op) and
   resend_invitation (fills the prior TODO) via BackgroundTasks. Email
   failures don't 5xx the invitation endpoint (try/except quiet send).

3. ACCESS_TOKEN_EXPIRE_HOURS env knob (api/security.py). Was hardcoded
   to 60 * 24. Now reads from env via int(float(os.getenv(...)) * 60)
   so fractional hours work — needed for the mobile/SMOKE.md refresh
   smoke (set to 0.05 ≈ 3 min, restart, walk smoke, revert). Settings
   type also widened to float for consistency. SMOKE.md refresh-test
   section now points at the env knob instead of the source-patch
   workaround.

Tests:
- tests/api/test_invitations_email.py: create + resend each queue the
  email with the right args + token; email-send-raises doesn't 5xx.
- tests/api/test_webhooks_sendgrid.py: missing sig → 400; bad sig →
  401; valid sig + each event type → expected Notification + DeliveryLog
  state; unknown sg_message_id dropped not 500; malformed JSON → 400.
- tests/api/test_access_token_expire_env.py: default 24h; integer
  hours; fractional hours (0.05 → 3 min). Module reloaded under env
  changes via importlib.reload.

Strikes 3 items from specs/024-sprint-9-completion/spec.md deferred list.

+++ api/core/config.py
@@ -23,7 +23,12 @@ class Settings(BaseSettings):
     # JWT Authentication
     SECRET_KEY: str = "change-this-to-a-random-secret-key-in-production"
     ALGORITHM: str = "HS256"
-    ACCESS_TOKEN_EXPIRE_HOURS: int = 24
+    # Float so sub-hour values work (e.g. 0.05 = 3 min for short-TTL
+    # refresh smoke per mobile/SMOKE.md). Read via os.getenv in
+    # api/security.py:24 — Settings is for code that needs structured
+    # access; the security module reads env directly to decouple from
+    # Settings reload cycles.
+    ACCESS_TOKEN_EXPIRE_HOURS: float = 24
 
     # Application
     APP_URL: str = "http://localhost:8000"
